fix protocol 776 clientbound codecs

+++ pkg/protocol/packet/game/client/sound_entity.go
@@ -13,5 +13,5 @@ type EntitySoundEffect struct {
 	EntityID      int32 `mc:"VarInt"`
 	Volume        float32
 	Pitch         float32
-	Seed          int32
+	Seed          int64
 }
